refactor(process-manager): move Sort out of the state aggregate
types.rs pulls nonos_libc for the procstat ABI, so nothing in it can be
compiled on the host. Sort is pure data that table_geom depends on, and the
geometry test has to reach it without the kernel ABI coming too.
